# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Finanças Pessoais API",
        "description": "API de controle financeiro pessoal com autenticação por token Sanctum e verificação de dois fatores (2FA).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor Principal"
        }
    ],
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Listar todas as categorias",
                "description": "Retorna a lista de categorias que pertencem ao usuário autenticado, ordenadas por nome.",
                "operationId": "ac9c4d978ea4e5b72915ca9728e5b8b1",
                "responses": {
                    "200": {
                        "description": "Lista de categorias.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid",
                                                        "example": "9c1c9ff5-5858-45bf-97c3-bc7948a4c84d"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Alimentação"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Supermercado e refeições"
                                                    },
                                                    "user_id": {
                                                        "type": "string",
                                                        "format": "uuid",
                                                        "example": "e129e061-0b86-4fb4-8975-d91d17983637"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Criar uma nova categoria",
                "description": "Cadastra uma categoria associada ao usuário autenticado.",
                "operationId": "cca49a87d9b0ebb9a98ad7c154b71737",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Transporte"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Combustível, Uber e passagens"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoria criada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Categoria criada com sucesso."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "a9ef52eb-5460-449e-b9ef-dcd41bb0b793"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Transporte"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Combustível, Uber e passagens"
                                                },
                                                "user_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "e129e061-0b86-4fb4-8975-d91d17983637"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado."
                    },
                    "422": {
                        "description": "Erro de validação (ex: nome faltando ou descrição muito curta)."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categories/{category}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Exibir detalhes de uma categoria",
                "description": "Retorna os dados de uma categoria específica. Apenas o proprietário da categoria pode acessá-la.",
                "operationId": "ff3ab984d3fe877eb16da6009394d694",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "UUID da categoria",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da categoria.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "a9ef52eb-5460-449e-b9ef-dcd41bb0b793"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Transporte"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Combustível, Uber e passagens"
                                                },
                                                "user_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "e129e061-0b86-4fb4-8975-d91d17983637"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado."
                    },
                    "403": {
                        "description": "Não autorizado (a categoria pertence a outro usuário)."
                    },
                    "404": {
                        "description": "Categoria não encontrada."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Atualizar uma categoria",
                "description": "Atualiza os campos de uma categoria existente pertencente ao usuário logado.",
                "operationId": "3146e96866d62a90bd01a7070fd61470",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "UUID da categoria",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Transporte e Viagens"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Gastos com transporte geral e viagens corporativas"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoria atualizada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Categoria atualizada com sucesso."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "a9ef52eb-5460-449e-b9ef-dcd41bb0b793"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Transporte e Viagens"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Gastos com transporte geral e viagens corporativas"
                                                },
                                                "user_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "e129e061-0b86-4fb4-8975-d91d17983637"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado."
                    },
                    "403": {
                        "description": "Não autorizado (a categoria pertence a outro usuário)."
                    },
                    "404": {
                        "description": "Categoria não encontrada."
                    },
                    "422": {
                        "description": "Erro de validação nos dados enviados."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Excluir uma categoria",
                "description": "Remove uma categoria pertencente ao usuário logado do sistema.",
                "operationId": "88c22abd8149aea4b3bdcb280a0fddf7",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "UUID da categoria",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoria excluída com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Categoria excluída com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado."
                    },
                    "403": {
                        "description": "Não autorizado (a categoria pertence a outro usuário)."
                    },
                    "404": {
                        "description": "Categoria não encontrada."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token recebido no fluxo de verificação 2FA (formato: Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}